draw_it v0.1.0 DrawIt.GameServer

A GenServer used for managing a game's state.

To create a game server, start by passing an existing %DrawIt.Games.Game{} into start_link/1.

{:ok, game} = Games.create_game(%{})
{:ok, server} = GameServer.start_link(game: game)

Once the server is created, you can use the returned server pid or the game's join_code to interact with the game server.

{:ok, player1} = GameServer.join(server, %{nickname: "Ben"})
{:ok, player2} = GameServer.join(game.join_code, %{nickname: "Wendy"})

{:ok, round} = GameServer.start_round(game.join_code, %{})
:ok = GameServer.end_round(game.join_code, %{})

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

end_round(join_code, payload)

guess(join_code, payload)

join(join_code, payload)

start_link(options)

start_round(join_code, payload)

whereis(join_code)